requirements.txt
numpy==1.23.5
pillow==9.4.0
matplotlib==3.7.1
albumentations==1.3.0
opencv-python-headless==4.7.0.72
seaborn==0.12.2
torch==2.0.1
torchvision==0.15.2
ipython==8.12.2
pytorch-ignite==0.4.11
tqdm==4.65.0
scikit-learn==1.2.2


